test(claim-spoofing): this commit contains the changes  required for existing test code to properly evaluate assertions against updated makeClaimAirdropInvitation
- added Observable factory function for watching AsyncGenerators
- implemented logic for informing smart wallets of tribbles issuer (fixes issue where provisionSmartWallet fails when interfacing with a new issuer)
- added import statements for various type definitions (a practice that should continue whenever faced with VSCode intellisense errors stemming from unknown types)
