+++ contract/test/tribbles-airdrop/contract.test.js
@@ -1,358 +1,641 @@
+/* eslint-disable import/order */
 // @ts-check
-
-/* eslint-disable import/order -- https://github.com/endojs/endo/issues/1235 */
+/* global setTimeout, fetch */
+// XXX what's the state-of-the-art in ava setup?
+// eslint-disable-next-line import/order
 import { test as anyTest } from '../prepare-test-env-ava.js';
 
 import { createRequire } from 'module';
-import { E } from '@endo/far';
-import { makeNodeBundleCache } from '@endo/bundle-source/cache.js';
-import { makeZoeKitForTest } from '@agoric/zoe/tools/setup-zoe.js';
-import { AmountMath } from '@agoric/ertp';
-
-import { makeStableFaucet } from '../mintStable.js';
-import buildManualTimer from '@agoric/zoe/tools/manualTimer.js';
-import { oneDay, TimeIntervals } from '../../src/helpers/time.js';
+import { env as ambientEnv } from 'node:process';
+import * as ambientChildProcess from 'node:child_process';
+import * as ambientFsp from 'node:fs/promises';
+import { E, passStyleOf } from '@endo/far';
 import { extract } from '@agoric/vats/src/core/utils.js';
+import {
+  makeTerms,
+  permit,
+  main,
+  startAirdrop,
+} from '../../src/airdrop.local.proposal.js';
+import {
+  makeBundleCacheContext,
+  getBundleId,
+} from '../../tools/bundle-tools.js';
+import { makeE2ETools } from '../../tools/e2e-tools.js';
+import {
+  makeNameProxy,
+  makeAgoricNames,
+} from '../../tools/ui-kit-goals/name-service-client.js';
 import { makeMockTools, mockBootstrapPowers } from '../../tools/boot-tools.js';
-import { getBundleId } from '../../tools/bundle-tools.js';
+import { merkleTreeAPI } from '../../src/merkle-tree/index.js';
+import { makeStableFaucet } from '../mintStable.js';
+import { makeOfferArgs } from './actors.js';
+import { merkleTreeObj } from './generated_keys.js';
+import { AmountMath } from '@agoric/ertp';
+import { Observable } from '../../src/helpers/adts.js';
+import { createStore } from '../../src/tribbles/utils.js';
 import { head } from '../../src/helpers/objectTools.js';
 
-import { simulateClaim } from './actors.js';
-import { OPEN } from '../../src/airdrop.contract.js';
-import {
-  startAirdrop,
-  permit,
-  makeTerms,
-} from '../../src/airdrop.local.proposal.js';
-import { makeFakeVatAdmin } from '@agoric/zoe/tools/fakeVatAdmin.js';
-import { merkleTreeObj } from './generated_keys.js';
+const reducerFn = (state = [], action) => {
+  const { type, payload } = action;
+  switch (type) {
+    case 'NEW_RESULT':
+      return [...state, payload];
+    default:
+      return state;
+  }
+};
+const handleNewResult = result => ({
+  type: 'NEW_RESULT',
+  payload: result.value,
+});
 
+const makeAsyncObserverObject = (
+  generator,
+  completeMessage = 'Iterator lifecycle complete.',
+  maxCount = Infinity,
+) =>
+  Observable(async observer => {
+    const iterator = E(generator);
+    const { dispatch, getStore } = createStore(reducerFn, []);
+    // eslint-disable-next-line no-constant-condition
+    while (true) {
+      // eslint-disable-next-line @jessie.js/safe-await-separator
+      const result = await iterator.next();
+      if (result.done) {
+        console.log('result.done === true #### breaking loop');
+        break;
+      }
+      dispatch(handleNewResult(result));
+      if (getStore().length === maxCount) {
+        console.log('getStore().length === maxCoutn');
+        break;
+      }
+      observer.next(result.value);
+    }
+    observer.complete({ message: completeMessage, values: getStore() });
+  });
+
+const traceFn = label => value => {
+  console.log(label, '::::', value);
+  return value;
+};
+
+const AIRDROP_TIERS_STATIC = [9000n, 6500n, 3500n, 1500n, 750n].map(
+  x => x * 1_000_000n,
+);
 const { accounts } = merkleTreeObj;
-/** @typedef {typeof import('../../src/airdrop.contract.js').start} AssetContractFn */
-
-const myRequire = createRequire(import.meta.url);
-const contractPath = myRequire.resolve(`../../src/airdrop.contract.js`);
-const AIRDROP_TIERS_STATIC = [9000n, 6500n, 3500n, 1500n, 750n];
+// import { makeAgdTools } from '../agd-tools.js';
 
 /** @type {import('ava').TestFn<Awaited<ReturnType<makeTestContext>>>} */
 const test = anyTest;
 
-const defaultCustomTerms = {
-  initialPayoutValues: AIRDROP_TIERS_STATIC,
-  targetNumberOfEpochs: 5,
-  targetEpochLength: TimeIntervals.SECONDS.ONE_DAY,
-  targetTokenSupply: 10_000_000n,
-  tokenName: 'Tribbles',
-  startTime: oneDay,
-  merkleRoot: merkleTreeObj.root,
+const nodeRequire = createRequire(import.meta.url);
+
+const bundleRoots = {
+  tribblesAirdrop: nodeRequire.resolve('../../src/airdrop.contract.js'),
 };
 
-const UNIT6 = 1_000_000n;
-
-const timerTracer = label => value => {
-  console.log(label, '::: latest #### ', value);
-  return value;
+const scriptRoots = {
+  tribblesAirdrop: nodeRequire.resolve('../../src/airdrop.local.proposal.js'),
 };
-const makeLocalTimer = async (
-  createTimerFn = buildManualTimer(timerTracer('default timer'), 5n),
-) => {
-  const timer = createTimerFn();
 
-  const timerBrand = await E(timer).getTimerBrand();
-
-  return {
-    timer,
-    timerBrand,
-  };
-};
-/**
- * Tests assume access to the zoe service and that contracts are bundled.
- *
- * See test-bundle-source.js for basic use of bundleSource().
- * Here we use a bundle cache to optimize running tests multiple times.
- *
- * @param {import('ava').TestFn} t
- *
- */
-
+/** @param {import('ava').ExecutionContext} t */
 const makeTestContext = async t => {
-  const { admin, vatAdminState } = makeFakeVatAdmin();
-  const { zoeService: zoe, feeMintAccess } = makeZoeKitForTest(admin);
+  const bc = await makeBundleCacheContext(t);
 
-  const invitationIssuer = zoe.getInvitationIssuer();
-  console.log('------------------------');
-  console.log('invitationIssuer::', invitationIssuer);
-  const bundleCache = await makeNodeBundleCache('bundles/', {}, s => import(s));
-  const bundle = await bundleCache.load(contractPath, 'assetContract');
-  const {} = makeMockTools(t, bundleCache);
+  const { E2E } = ambientEnv;
+  const { execFileSync, execFile } = ambientChildProcess;
+  const { writeFile } = ambientFsp;
 
-  const testFeeIssuer = await E(zoe).getFeeIssuer();
-  const testFeeBrand = await E(testFeeIssuer).getBrand();
-
-  const testFeeTokenFaucet = await makeStableFaucet({
-    feeMintAccess,
-    zoe,
-    bundleCache,
-  });
-  console.log('bundle:::', { bundle, bundleCache });
-  return {
-    invitationIssuer,
-    zoe,
-    bundle,
-    bundleCache,
-    makeLocalTimer,
-    testFeeTokenFaucet,
-    faucet: testFeeTokenFaucet.faucet,
-    testFeeBrand,
-    testFeeIssuer,
+  /** @type {import('../../tools/agd-lib.js').ExecSync} */
+  const dockerExec = (file, args, opts = { encoding: 'utf-8' }) => {
+    const workdir = '/workspace/contract';
+    const execArgs = ['compose', 'exec', '--workdir', workdir, 'agd'];
+    opts.verbose &&
+      console.log('docker compose exec', JSON.stringify([file, ...args]));
+    return execFileSync('docker', [...execArgs, file, ...args], opts);
   };
+
+  console.time('makeTestTools');
+  console.timeLog('makeTestTools', 'start');
+  // installBundles,
+  // runCoreEval,
+  // provisionSmartWallet,
+  // runPackageScript???
+  const tools = await (E2E
+    ? makeE2ETools(t, bc.bundleCache, {
+        execFileSync: dockerExec,
+        execFile,
+        fetch,
+        setTimeout,
+        writeFile,
+      })
+    : makeMockTools(t, bc.bundleCache));
+  console.timeEnd('makeTestTools');
+
+  return { ...tools, ...bc };
 };
 
-test.before('before hook', async t => {
-  const { admin, vatAdminState } = makeFakeVatAdmin();
-  const { zoeService: zoe, feeMintAccess } = makeZoeKitForTest(admin);
+test.before(async t => (t.context = await makeTestContext(t)));
 
-  const invitationIssuer = zoe.getInvitationIssuer();
-  console.log('------------------------');
-  console.log('invitationIssuer::', invitationIssuer);
-  const bundleCache = await makeNodeBundleCache('bundles/', {}, s => import(s));
-  const bundle = await bundleCache.load(contractPath, 'assetContract');
-  const {} = makeMockTools(t, bundleCache);
+//  console.log('after makeAgdTools:::', { context: t.context });
 
-  const testFeeIssuer = await E(zoe).getFeeIssuer();
-  const testFeeBrand = await E(testFeeIssuer).getBrand();
-
-  const testFeeTokenFaucet = await makeStableFaucet({
-    feeMintAccess,
-    zoe,
-    bundleCache,
-  });
-  console.log('bundle:::', { bundle, bundleCache });
-  t.context = {
-    invitationIssuer,
-    zoe,
-    bundle,
-    bundleCache,
-    makeLocalTimer,
-    testFeeTokenFaucet,
-    faucet: testFeeTokenFaucet.faucet,
-    testFeeBrand,
-    testFeeIssuer,
+test.serial('we1ll-known brand (ATOM) is available', async t => {
+  const { makeQueryTool } = t.context;
+  const hub0 = makeAgoricNames(makeQueryTool());
+  const agoricNames = makeNameProxy(hub0);
+  await null;
+  const brand = {
+    ATOM: await agoricNames.brand.ATOM,
   };
+  t.log(brand);
+  t.is(passStyleOf(brand.ATOM), 'remotable');
 });
 
-// IDEA: use test.serial and pass work products
-// between tests using t.context.
-
-test('Install the contract', async t => {
-  const { zoe, bundle } = t.context;
-
-  const installation = await E(zoe).install(bundle);
-  t.log(installation);
-  t.is(typeof installation, 'object');
-});
-
-const startLocalInstance = async (
-  t,
-  bundle,
-  { issuers: { Fee: feeIssuer }, zoe, terms: customTerms },
-) => {
-  const timer = buildManualTimer();
-
-  /** @type {ERef<Installation<AssetContractFn>>} */
-  const installation = await E(zoe).install(bundle);
-
-  const { instance, publicFacet, creatorFacet } = await E(zoe).startInstance(
-    installation,
-    { Fee: feeIssuer },
-    {
-      ...customTerms,
-    },
-    { timer },
+test.serial('install bundle: airdrop / tribblesAirdrop', async t => {
+  const { installBundles } = t.context;
+  console.time('installBundles');
+  console.timeLog('installBundles', Object.keys(bundleRoots).length, 'todo');
+  const bundles = await installBundles(bundleRoots, (...args) =>
+    console.timeLog('installBundles', ...args),
   );
 
-  t.log('instance', { instance });
+  console.timeEnd('installBundles');
 
-  return { instance, installation, timer, publicFacet, creatorFacet };
-};
-
-test.serial('Start the contract', async t => {
-  const {
-    zoe: zoeRef,
-    bundle,
-    bundleCache,
-    feeMintAccess,
-    testFeeBrand,
-  } = t.context;
-
-  const testFeeIssuer = await E(zoeRef).getFeeIssuer();
-
-  const testFeeTokenFaucet = await makeStableFaucet({
-    feeMintAccess,
-    zoe: zoeRef,
-    bundleCache,
-  });
-  console.log('context:', { testFeeTokenFaucet });
-
-  const localTestConfig = {
-    zoe: zoeRef,
-    issuers: { Fee: testFeeIssuer },
-    terms: {
-      ...defaultCustomTerms,
-      feeAmount: AmountMath.make(testFeeBrand, 5n),
-    },
-  };
-
-  const { instance } = await startLocalInstance(t, bundle, localTestConfig);
-  t.log(instance);
-  t.is(typeof instance, 'object');
-});
-
-test('Airdrop ::: happy paths', async t => {
-  const { zoe: zoeRef, bundle, faucet, testFeeBrand } = await t.context;
-  console.log(t.context);
-  const { instance, publicFacet, timer } = await startLocalInstance(t, bundle, {
-    zoe: zoeRef,
-    issuers: { Fee: await E(zoeRef).getFeeIssuer() },
-    terms: {
-      ...defaultCustomTerms,
-      feeAmount: AmountMath.make(testFeeBrand, 5n),
-    },
-  });
-
-  await E(timer).advanceBy(oneDay * (oneDay / 2n));
-
-  t.deepEqual(await E(publicFacet).getStatus(), OPEN);
-
-  t.log({ faucet });
-  await E(timer).advanceBy(oneDay);
-  const feePurse = await faucet(5n * UNIT6);
-  t.log(feePurse);
-  await t.deepEqual(
-    feePurse.getCurrentAmount(),
-    AmountMath.make(t.context.testFeeBrand, 5_000_000n),
+  const id = getBundleId(bundles.tribblesAirdrop);
+  const shortId = id.slice(0, 8);
+  t.log('bundleId', shortId);
+  t.is(id.length, 3 + 128, 'bundleID length');
+  t.regex(id, /^b1-.../);
+  console.groupEnd();
+  Object.assign(t.context.shared, { bundles });
+  t.truthy(
+    t.context.shared.bundles.tribblesAirdrop,
+    't.context.shared.bundles should contain a property "tribblesAirdrop"',
   );
-
-  await simulateClaim(t, zoeRef, instance, feePurse, head(accounts));
-
-  await E(timer).advanceBy(oneDay);
-
-  await simulateClaim(t, zoeRef, instance, feePurse, accounts[2]);
-
-  await E(timer).advanceBy(oneDay);
-
-  t.deepEqual(await E(publicFacet).getStatus(), 'claim-window-open');
-
-  await E(timer).advanceBy(oneDay);
 });
+const containsSubstring = (substring, string) =>
+  new RegExp(substring, 'i').test(string);
 
-test.serial('delegate pause access :: makePauseContractInvitation', async t => {
-  const {
-    zoe: zoeRef,
-    invitationIssuer: zoeIssuer,
-    bundle,
-    faucet,
-    testFeeBrand,
-  } = await t.context;
-  console.log(t.context);
-
-  const invitationPurse = await E(zoeIssuer).makeEmptyPurse();
-  const depositOnlyFacet = invitationPurse.getDepositFacet();
-
-  const { instance, publicFacet, timer, creatorFacet } =
-    await startLocalInstance(t, bundle, {
-      zoe: zoeRef,
-      issuers: { Fee: await E(zoeRef).getFeeIssuer() },
-      terms: {
-        ...defaultCustomTerms,
-        feeAmount: AmountMath.make(testFeeBrand, 5n),
-      },
-    });
-
-  await E(creatorFacet).makePauseContractInvitation(depositOnlyFacet);
-
-  const pauseInvitationAmt = invitationPurse.getCurrentAmount();
+test.serial('deploy contract with core eval: airdrop / airdrop', async t => {
+  const { runCoreEval } = t.context;
+  const { bundles } = t.context.shared;
+  const bundleID = getBundleId(bundles.tribblesAirdrop);
 
   t.deepEqual(
-    pauseInvitationAmt.brand,
-    await E(zoeIssuer).getBrand(),
-    'makePauseContractInvitation given a valid depositFacet should deposit an invitation into its purse.',
-  );
-
-  const pauseOffersPayment = invitationPurse.withdraw(pauseInvitationAmt);
-
-  // Claming is not yet active in contract.
-  // Code below produces: "Illegal state transition. Can not transition from state: prepared to state paused."
-  //  await E(zoeRef).offer(pauseOffersPayment, undefined, undefined);
-
-  await E(timer).advanceBy(oneDay * (oneDay / 2n));
-
-  t.deepEqual(await E(publicFacet).getStatus(), OPEN);
-
-  await E(timer).advanceBy(oneDay);
-  const feePurse = await faucet(5n * UNIT6);
-
-  await simulateClaim(t, zoeRef, instance, feePurse, accounts[2]);
-
-  const adminSeat = await E(zoeRef).offer(
-    pauseOffersPayment,
-    undefined,
-    undefined,
-  );
-
-  t.deepEqual(
-    await adminSeat.hasExited(),
+    containsSubstring(bundles.tribblesAirdrop.endoZipBase64Sha512, bundleID),
     true,
-    'adminSeat.hasExited() should return true following a succesful offer to pause the contract.',
+  );
+  const merkleRoot = merkleTreeAPI.generateMerkleRoot(
+    accounts.map(x => x.pubkey.key),
   );
 
-  await E(timer).advanceBy(oneDay);
+  const { vatAdminState } = await mockBootstrapPowers(t.log);
 
-  await E(timer).advanceBy(oneDay);
+  vatAdminState.installBundle(bundleID, bundles.tribblesAirdrop);
 
-  t.deepEqual(await E(publicFacet).getStatus(), 'paused');
-
-  // TODO: Validate that an offer make to contract fails when offer filter is present
-});
-
-test.serial(
-  'MN-2 Task: Add a deployment test that exercises the core-eval that will be used to install & start the contract on chain.',
-  async t => {
-    const { bundle, testFeeBrand } = t.context;
-
-    const bundleID = getBundleId(bundle);
-    const { powers, vatAdminState } = await mockBootstrapPowers(t.log);
-    const { feeMintAccess, zoe } = powers.consume;
-
-    // When the BLD staker governance proposal passes,
-    // the startup function gets called.
-    vatAdminState.installBundle(bundleID, bundle);
-    const airdropPowers = extract(permit, powers);
-    await startAirdrop(airdropPowers, {
-      merkleRoot: merkleTreeObj.root,
+  console.log('inside deploy test::', bundleID);
+  // this runCoreEval does not work
+  const name = 'airdrop';
+  const result = await runCoreEval({
+    name,
+    behavior: main,
+    entryFile: scriptRoots.tribblesAirdrop,
+    config: {
       options: {
         customTerms: {
           ...makeTerms(),
           merkleRoot: merkleTreeObj.root,
         },
         tribblesAirdrop: { bundleID },
+        merkleRoot,
+      },
+    },
+  });
+
+  t.log(result.voting_end_time, '#', result.proposal_id, name);
+  t.like(result, {
+    content: {
+      '@type': '/agoric.swingset.CoreEvalProposal',
+    },
+    status: 'PROPOSAL_STATUS_PASSED',
+  });
+});
+
+const makeMakeOfferSpec = instance => (account, feeAmount, id) => ({
+  id: `offer-${id}`,
+  invitationSpec: {
+    source: 'contract',
+    instance,
+    publicInvitationMaker: 'makeClaimTokensInvitation',
+  },
+  proposal: { give: { Fee: feeAmount } },
+  offerArgs: { ...makeOfferArgs(account) },
+});
+test.serial('E2E test', async t => {
+  const merkleRoot = merkleTreeAPI.generateMerkleRoot(
+    accounts.map(x => x.pubkey.key),
+  );
+  const { bundleCache } = t.context;
+
+  t.log('starting contract with merkleRoot:', merkleRoot);
+  // Is there a better way to obtain a reference to this bundle???
+  // or is this just fine??
+  const { tribblesAirdrop } = t.context.shared.bundles;
+
+  const bundleID = getBundleId(tribblesAirdrop);
+  const { powers, vatAdminState, makeMockWalletFactory, provisionSmartWallet } =
+    await makeMockTools(t, bundleCache);
+  const { feeMintAccess, zoe } = powers.consume;
+
+  vatAdminState.installBundle(bundleID, tribblesAirdrop);
+  const adminWallet = await provisionSmartWallet(
+    'agoric1jng25adrtpl53eh50q7fch34e0vn4g72j6zcml',
+    {
+      BLD: 10n,
+    },
+  );
+
+  const zoeIssuer = await E(zoe).getInvitationIssuer();
+
+  const zoeBrand = await zoeIssuer.getBrand();
+  const adminZoePurse = E(adminWallet.peek).purseUpdates(zoeBrand);
+
+  const airdropPowers = extract(permit, powers);
+
+  await startAirdrop(airdropPowers, {
+    options: {
+      customTerms: {
+        ...makeTerms(),
         merkleRoot: merkleTreeObj.root,
       },
-    });
-    const sellSpace = powers;
-    const instance = await sellSpace.instance.consume.tribblesAirdrop;
-    console.log({ powers });
+      tribblesAirdrop: { bundleID },
+    },
+  });
 
-    // Now that we have the instance, resume testing as above.
-    const { bundleCache } = t.context;
-    const { faucet } = makeStableFaucet({ bundleCache, feeMintAccess, zoe });
+  await makeAsyncObserverObject(
+    adminZoePurse,
+    'invitation recieved',
+    1,
+  ).subscribe({
+    next: traceFn('ADMIN_WALLET::: NEXT'),
+    error: traceFn('ADMIN WALLET::: ERROR'),
+    complete: async ({ message, values }) => {
+      const [pauseInvitationDetails] = values;
+      t.deepEqual(message, 'invitation recieved');
+      t.deepEqual(pauseInvitationDetails.brand, zoeBrand);
+      t.deepEqual(
+        head(pauseInvitationDetails.value).description,
+        'pause contract',
+      );
+    },
+  });
+  /** @type {import('../../src/airdrop.local.proposal.js').AirdropSpace} */
+  // @ts-expeimport { merkleTreeObj } from '@agoric/orchestration/src/examples/airdrop/generated_keys.js';
+  const airdropSpace = powers;
+  const instance = await airdropSpace.instance.consume.tribblesAirdrop;
 
-    await simulateClaim(
-      t,
-      zoe,
+  const terms = await E(zoe).getTerms(instance);
+
+  const { issuers, brands } = terms;
+
+  const walletFactory = makeMockWalletFactory({
+    Tribbles: issuers.Tribbles,
+    Fee: issuers.Fee,
+  });
+
+  const wallets = {
+    alice: await walletFactory.makeSmartWallet(accounts[4].address),
+    bob: await walletFactory.makeSmartWallet(accounts[2].address),
+  };
+  const { faucet, mintBrandedPayment } = makeStableFaucet({
+    bundleCache,
+    feeMintAccess,
+    zoe,
+  });
+
+  await Object.values(wallets).map(async wallet => {
+    const pmt = await mintBrandedPayment(10n);
+    console.log('payment::', pmt);
+    await E(wallet.deposit).receive(pmt);
+  });
+  const makeOfferSpec = makeMakeOfferSpec(instance);
+
+  await faucet(5n * 1_000_000n);
+
+  const makeFeeAmount = () => AmountMath.make(brands.Fee, 5n);
+
+  const [aliceTier, bobTier] = [0, 2];
+  const [alice, bob] = [
+    [
+      E(wallets.alice.offers).executeOffer(
+        makeOfferSpec({ ...accounts[4], tier: 0 }, makeFeeAmount(), 0),
+      ),
+      E(wallets.alice.peek).purseUpdates(brands.Tribbles),
+    ],
+    [
+      E(wallets.bob.offers).executeOffer(
+        makeOfferSpec({ ...accounts[2], tier: bobTier }, makeFeeAmount(), 0),
+      ),
+      E(wallets.bob.peek).purseUpdates(brands.Tribbles),
+    ],
+  ];
+
+  const [alicesOfferUpdates, alicesPurse] = alice;
+  const [bobsOfferUpdate, bobsPurse] = bob;
+  /**
+   * @typedef {{value: { updated: string, status: { id: string, invitationSpec: import('../../tools/wallet-tools.js').InvitationSpec, proposal:Proposal, offerArgs: {key: string, proof: []}}}}} OfferResult
+   */
+
+  await makeAsyncObserverObject(alicesOfferUpdates).subscribe({
+    next: traceFn('SUBSCRIBE.NEXT'),
+    error: traceFn('AliceOffer Error'),
+    complete: ({ message, values }) => {
+      t.deepEqual(message, 'Iterator lifecycle complete.');
+      t.deepEqual(values.length, 4);
+    },
+  });
+
+  await makeAsyncObserverObject(
+    alicesPurse,
+    'AsyncGenerator alicePurse has fufilled its requirements.',
+    1,
+  ).subscribe({
+    next: traceFn('TRIBBLES_WATCHER ### SUBSCRIBE.NEXT'),
+    error: traceFn('TRIBBLES_WATCHER #### SUBSCRIBE.ERROR'),
+    complete: ({ message, values }) => {
+      t.deepEqual(
+        message,
+        'AsyncGenerator alicePurse has fufilled its requirements.',
+      );
+      t.deepEqual(
+        head(values),
+        AmountMath.make(brands.Tribbles, AIRDROP_TIERS_STATIC[aliceTier]),
+      );
+    },
+  });
+  const [alicesSecondClaim] = [
+    E(wallets.alice.offers).executeOffer(
+      makeOfferSpec({ ...accounts[4], tier: 0 }, makeFeeAmount(), 0),
+    ),
+  ];
+
+  const alicesSecondOfferSubscriber = makeAsyncObserverObject(
+    alicesSecondClaim,
+  ).subscribe({
+    next: traceFn('alicesSecondClaim ### SUBSCRIBE.NEXT'),
+    error: traceFn('alicesSecondClaim #### SUBSCRIBE.ERROR'),
+    complete: traceFn('alicesSecondClaim ### SUBSCRIBE.COMPLETE'),
+  });
+  await t.throwsAsync(alicesSecondOfferSubscriber, {
+    message: `Allocation for address ${accounts[4].address} has already been claimed.`,
+  });
+
+  await makeAsyncObserverObject(
+    bobsOfferUpdate,
+    'AsyncGenerator bobsOfferUpdate has fufilled its requirements.',
+  ).subscribe({
+    next: traceFn('BOBS_OFFER_UPDATE:::: SUBSCRIBE.NEXT'),
+    error: traceFn('BOBS_OFFER_UPDATE:::: SUBSCRIBE.ERROR'),
+    complete: ({ message, values }) => {
+      t.deepEqual(
+        message,
+        'AsyncGenerator bobsOfferUpdate has fufilled its requirements.',
+      );
+      t.deepEqual(values.length, 4);
+    },
+  });
+
+  await makeAsyncObserverObject(
+    bobsPurse,
+    'AsyncGenerator bobsPurse has fufilled its requirements.',
+    1,
+  ).subscribe({
+    next: traceFn('TRIBBLES_WATCHER ### SUBSCRIBE.NEXT'),
+    error: traceFn('TRIBBLES_WATCHER #### SUBSCRIBE.ERROR'),
+    complete: ({ message, values }) => {
+      t.deepEqual(
+        message,
+        'AsyncGenerator bobsPurse has fufilled its requirements.',
+      );
+      t.deepEqual(
+        head(values),
+        AmountMath.make(brands.Tribbles, AIRDROP_TIERS_STATIC[bobTier]),
+      );
+    },
+  });
+});
+test.serial('makePauseContractInvitation', async t => {
+  const merkleRoot = merkleTreeAPI.generateMerkleRoot(
+    accounts.map(x => x.pubkey.key),
+  );
+  const { bundleCache } = t.context;
+
+  t.log('starting contract with merkleRoot:', merkleRoot);
+  // Is there a better way to obtain a reference to this bundle???
+  // or is this just fine??
+  const { tribblesAirdrop } = t.context.shared.bundles;
+
+  const bundleID = getBundleId(tribblesAirdrop);
+  const { powers, vatAdminState, makeMockWalletFactory, provisionSmartWallet } =
+    await makeMockTools(t, bundleCache);
+  const { feeMintAccess, zoe } = powers.consume;
+
+  vatAdminState.installBundle(bundleID, tribblesAirdrop);
+  const adminWallet = await provisionSmartWallet(
+    'agoric1jng25adrtpl53eh50q7fch34e0vn4g72j6zcml',
+    {
+      BLD: 10n,
+    },
+  );
+
+  const zoeIssuer = await E(zoe).getInvitationIssuer();
+
+  const zoeBrand = await zoeIssuer.getBrand();
+  const adminZoePurse = E(adminWallet.peek).purseUpdates(zoeBrand);
+
+  const airdropPowers = extract(permit, powers);
+
+  await startAirdrop(airdropPowers, {
+    options: {
+      customTerms: {
+        ...makeTerms(),
+        merkleRoot: merkleTreeObj.root,
+      },
+      tribblesAirdrop: { bundleID },
+    },
+  });
+
+  await makeAsyncObserverObject(
+    adminZoePurse,
+    'invitation recieved',
+    1,
+  ).subscribe({
+    next: traceFn('ADMIN_WALLET::: NEXT'),
+    error: traceFn('ADMIN WALLET::: ERROR'),
+    complete: async ({ message, values }) => {
+      const [pauseInvitationDetails] = values;
+      t.deepEqual(message, 'invitation recieved');
+      t.deepEqual(pauseInvitationDetails.brand, zoeBrand);
+      t.deepEqual(
+        head(pauseInvitationDetails.value).description,
+        'pause contract',
+      );
+    },
+  });
+  /** @type {import('../../src/airdrop.local.proposal.js').AirdropSpace} */
+  // @ts-expeimport { merkleTreeObj } from '@agoric/orchestration/src/examples/airdrop/generated_keys.js';
+  const airdropSpace = powers;
+  const instance = await airdropSpace.instance.consume.tribblesAirdrop;
+
+  const terms = await E(zoe).getTerms(instance);
+  const { issuers, brands } = terms;
+
+  const walletFactory = makeMockWalletFactory({
+    Tribbles: issuers.Tribbles,
+    Fee: issuers.Fee,
+  });
+
+  const wallets = {
+    alice: await walletFactory.makeSmartWallet(accounts[4].address),
+    bob: await walletFactory.makeSmartWallet(accounts[2].address),
+  };
+  const { faucet, mintBrandedPayment } = makeStableFaucet({
+    bundleCache,
+    feeMintAccess,
+    zoe,
+  });
+
+  await Object.values(wallets).map(async wallet => {
+    const pmt = await mintBrandedPayment(10n);
+    console.log('payment::', pmt);
+    await E(wallet.deposit).receive(pmt);
+  });
+  const makeOfferSpec = makeMakeOfferSpec(instance);
+
+  await faucet(5n * 1_000_000n);
+
+  const makeFeeAmount = () => AmountMath.make(brands.Fee, 5n);
+
+  const pauseOffer = {
+    id: 'admin-pause-1',
+    invitationSpec: {
+      source: 'purse',
       instance,
-      await faucet(5n * UNIT6),
-      accounts[3],
-    );
-  },
-);
+      description: 'pause contract',
+    },
+    proposal: {},
+  };
+  await E(adminWallet.offers).executeOffer(pauseOffer);
+  const [aliceTier, bobTier] = [0, 2];
+  const [alice, bob] = [
+    [
+      E(wallets.alice.offers).executeOffer(
+        makeOfferSpec({ ...accounts[4], tier: 0 }, makeFeeAmount(), 0),
+      ),
+      E(wallets.alice.peek).purseUpdates(brands.Tribbles),
+    ],
+    [
+      E(wallets.bob.offers).executeOffer(
+        makeOfferSpec({ ...accounts[2], tier: bobTier }, makeFeeAmount(), 0),
+      ),
+      E(wallets.bob.peek).purseUpdates(brands.Tribbles),
+    ],
+  ];
+
+  const [alicesOfferUpdates, alicesPurse] = alice;
+  const [bobsOfferUpdate, bobsPurse] = bob;
+  /**
+   * @typedef {{value: { updated: string, status: { id: string, invitationSpec: import('../../tools/wallet-tools.js').InvitationSpec, proposal:Proposal, offerArgs: {key: string, proof: []}}}}} OfferResult
+   */
+
+  await makeAsyncObserverObject(alicesOfferUpdates).subscribe({
+    next: traceFn('AliceOffer.NEXT ## AFTER PAUSE'),
+    error: traceFn('AliceOffer Error'),
+    complete: ({ message, values }) => {
+      t.deepEqual(message, 'Iterator lifecycle complete.');
+      t.deepEqual(values.length, 4);
+    },
+  });
+
+  await makeAsyncObserverObject(
+    alicesPurse,
+    'AsyncGenerator alicePurse has fufilled its requirements.',
+    1,
+  ).subscribe({
+    next: traceFn('TRIBBLES_WATCHER ### SUBSCRIBE.NEXT'),
+    error: traceFn('TRIBBLES_WATCHER #### SUBSCRIBE.ERROR'),
+    complete: ({ message, values }) => {
+      t.deepEqual(
+        message,
+        'AsyncGenerator alicePurse has fufilled its requirements.',
+      );
+      t.deepEqual(
+        head(values),
+        AmountMath.make(brands.Tribbles, AIRDROP_TIERS_STATIC[aliceTier]),
+      );
+    },
+  });
+  const [alicesSecondClaim] = [
+    E(wallets.alice.offers).executeOffer(
+      makeOfferSpec({ ...accounts[4], tier: 0 }, makeFeeAmount(), 0),
+    ),
+  ];
+
+  const alicesSecondOfferSubscriber = makeAsyncObserverObject(
+    alicesSecondClaim,
+  ).subscribe({
+    next: traceFn('alicesSecondClaim ### SUBSCRIBE.NEXT'),
+    error: traceFn('alicesSecondClaim #### SUBSCRIBE.ERROR'),
+    complete: traceFn('alicesSecondClaim ### SUBSCRIBE.COMPLETE'),
+  });
+  await t.throwsAsync(alicesSecondOfferSubscriber, {
+    message: `Allocation for address ${accounts[4].address} has already been claimed.`,
+  });
+
+  await makeAsyncObserverObject(
+    bobsOfferUpdate,
+    'AsyncGenerator bobsOfferUpdate has fufilled its requirements.',
+  ).subscribe({
+    next: traceFn('BOBS_OFFER_UPDATE:::: SUBSCRIBE.NEXT'),
+    error: traceFn('BOBS_OFFER_UPDATE:::: SUBSCRIBE.ERROR'),
+    complete: ({ message, values }) => {
+      t.deepEqual(
+        message,
+        'AsyncGenerator bobsOfferUpdate has fufilled its requirements.',
+      );
+      t.deepEqual(values.length, 4);
+    },
+  });
+
+  await makeAsyncObserverObject(
+    bobsPurse,
+    'AsyncGenerator bobsPurse has fufilled its requirements.',
+    1,
+  ).subscribe({
+    next: traceFn('TRIBBLES_WATCHER ### SUBSCRIBE.NEXT'),
+    error: traceFn('TRIBBLES_WATCHER #### SUBSCRIBE.ERROR'),
+    complete: ({ message, values }) => {
+      t.deepEqual(
+        message,
+        'AsyncGenerator bobsPurse has fufilled its requirements.',
+      );
+      t.deepEqual(
+        head(values),
+        AmountMath.make(brands.Tribbles, AIRDROP_TIERS_STATIC[bobTier]),
+      );
+    },
+  });
+});
+test.serial('agoricNames.instances has contract: airdrop', async t => {
+  const { makeQueryTool } = t.context;
+  const hub0 = makeAgoricNames(makeQueryTool());
+
+  const agoricNames = makeNameProxy(hub0);
+  console.log({ agoricNames });
+  await null;
+  const instance = await E(agoricNames).lookup('instance', 'tribblesAirdrop');
+  t.is(passStyleOf(instance), 'remotable');
+  t.log(instance);
+});
