+++ contract/test/types.js
@@ -1,0 +1,9 @@
+/** @import { Brand, Issuer } from '@agoric/ertp/src/types.js'; */
+/**
+ * @typedef {{
+ *   brand: Record<string, Promise<Brand>> & { timer: unknown }
+ *   issuer: Record<string, Promise<Issuer>>
+ *   instance: Record<string, Promise<Instance>>
+ *   installation: Record<string, Promise<Installation>>
+ * }} WellKnown
+ */
